Update VoltageReferenceProgrammableTeensyOLEDrev300.ino
Changed MAXVAL and MINVAL from uint32_t to int16_t integers to correct:
warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
if (refVal > MAXVAL) refVal = 1;

+++ VoltageReferenceProgrammableTeensyOLEDrev300/VoltageReferenceProgrammableTeensyOLEDrev300.ino
@@ -74,8 +74,8 @@ Bounce bm = Bounce();
 JC_EEPROM eep(JC_EEPROM::kbits_512, 2, 64);      //one 24LC512 EEPROM on the bus - device size, number of devices, page size
   
 // Instantiate objects used in this project
-uint32_t MAXVAL = 5000;    // DAC Maximum value
-uint32_t MINVAL = 1;        // DAC Minimum value
+int16_t MAXVAL = 5000;    // DAC Maximum value
+int16_t MINVAL = 1;        // DAC Minimum value
 
 // LED colors used in LEDcolor()
 uint8_t red = 0x1;
